ShortestPath/requirements.txt
dronekit==2.9.2
rplidar==0.9.2
paho-mqtt==1.6.1
pyttsx3==2.90
playaudio==1.0.1
pyserial==3.5
